+++ pkg/controller/cronbackupcontroller/controller.go
@@ -23,6 +23,7 @@ import (
 	"fmt"
 	"sort"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/robfig/cron/v3"
@@ -36,7 +37,6 @@ import (
 	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/util/rand"
 
-	v1core "github.com/kubeclipper/kubeclipper/pkg/apis/core/v1"
 	"github.com/kubeclipper/kubeclipper/pkg/client/informers"
 	listerv1 "github.com/kubeclipper/kubeclipper/pkg/client/lister/core/v1"
 	"github.com/kubeclipper/kubeclipper/pkg/component"
@@ -128,7 +128,7 @@ func (r *CronBackupReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 	if cronBackup.Spec.Schedule != "" {
 		// first created cron backup, update next schedule time
 		if cronBackup.Status.NextScheduleTime == nil {
-			schedule := v1core.ParseSchedule(cronBackup.Spec.Schedule)
+			schedule := parseSchedule(cronBackup.Spec.Schedule)
 			s, _ := cron.NewParser(4 | 8 | 16 | 32 | 64).Parse(schedule)
 			// update the next schedule time
 			nextRunAt := metav1.NewTime(s.Next(time.Now()))
@@ -155,7 +155,7 @@ func (r *CronBackupReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 			cronBackup.Status.LastScheduleTime = &now
 			cronBackup.Status.LastSuccessfulTime = cronBackup.Status.LastScheduleTime
 
-			schedule := v1core.ParseSchedule(cronBackup.Spec.Schedule)
+			schedule := parseSchedule(cronBackup.Spec.Schedule)
 			s, _ := cron.NewParser(4 | 8 | 16 | 32 | 64).Parse(schedule)
 			// update the next schedule time
 			nextRunAt := metav1.NewTime(s.Next(now.Time))
@@ -535,4 +535,23 @@ func (r *CronBackupReconciler) deleteBackup(log logger.Logging, clusterName stri
 		log.Error("Failed to delivery operation", zap.Error(err))
 	}()
 	return nil
+}
+
+func parseSchedule(schedule string) string {
+	year := time.Now().Year()
+	arr := strings.Split(schedule, " ")
+	if arr[2] == "L" {
+		switch time.Now().Month() {
+		case 1, 3, 5, 7, 8, 10, 12:
+			return strings.Replace(schedule, "L", "31", 1)
+		case 4, 6, 9, 11:
+			return strings.Replace(schedule, "L", "30", 1)
+		case 2:
+			if (year%4 == 0 && year%100 != 0) || year%400 == 0 {
+				return strings.Replace(schedule, "L", "28", 1)
+			}
+			return strings.Replace(schedule, "L", "29", 1)
+		}
+	}
+	return schedule
 }
